requirements.txt
#!/usr/bin/env bash

python-dotenv
apiai
sshtunnel
flask